♻️ 리팩토링 : CocktailController의 구조를 수정
CocktailSearch 내부에 있던 enum을 외부로 분리함.
  - AlcoholType
  - SearchType
CustomCocktailRepositoryImpl을 분리한 enum을 참조하도록 수정
