+++ src/main/java/org/toastit_v2/feature/basecocktail/domain/AlcoholType.java
@@ -1,0 +1,11 @@
+package org.toastit_v2.feature.basecocktail.domain;
+
+public enum AlcoholType {
+
+    /**
+     * AlcoholType
+     */
+    ALCOHOLIC,      // 알코올
+    NON_ALCOHOLIC   // 논알코올
+
+}
